test-requirements.txt
"""Test requirements file."""
# Testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-xdist>=3.0.0

# Code quality
flake8>=6.0.0
mypy>=1.0.0
black>=23.0.0
isort>=5.12.0

# Performance testing
pytest-benchmark>=4.0.0
memory-profiler>=0.60.0

# HTTP testing
requests-mock>=1.10.0
responses>=0.23.0

# Flask testing
flask-testing>=0.8.1

# Coverage reporting
coverage>=7.0.0
codecov>=2.1.0
